+++ packages/reown_appkit/example/base/lib/pages/settings_page.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:reown_appkit/reown_appkit.dart';
 
+// ignore: depend_on_referenced_packages
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:reown_appkit_dapp/utils/constants.dart';
 
 class SettingsPage extends StatefulWidget {
@@ -35,15 +37,11 @@ class SettingsPageState extends State<SettingsPage> {
       fontSize: 12.0,
       color: ReownAppKitModalTheme.colorsOf(context).foreground100,
     );
-    final textStyleBold = textStyle.copyWith(
-      fontWeight: FontWeight.bold,
-    );
+    final textStyleBold = textStyle.copyWith(fontWeight: FontWeight.bold);
     final redirect = widget.appKitModal.appKit!.metadata.redirect;
     return Center(
       child: Container(
-        constraints: const BoxConstraints(
-          maxWidth: StyleConstants.maxWidth,
-        ),
+        constraints: const BoxConstraints(maxWidth: StyleConstants.maxWidth),
         padding: const EdgeInsets.only(left: 12.0, right: 12.0),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
@@ -60,8 +58,9 @@ class SettingsPageState extends State<SettingsPage> {
                             'Relay Mode\n(Multichain)',
                             textAlign: TextAlign.end,
                             style: TextStyle(
-                              color: ReownAppKitModalTheme.colorsOf(context)
-                                  .foreground100,
+                              color: ReownAppKitModalTheme.colorsOf(
+                                context,
+                              ).foreground100,
                               fontWeight: !widget.linkMode
                                   ? FontWeight.bold
                                   : FontWeight.normal,
@@ -88,8 +87,9 @@ class SettingsPageState extends State<SettingsPage> {
                           child: Text(
                             'Link Mode\n(1CA, only EVM)',
                             style: TextStyle(
-                              color: ReownAppKitModalTheme.colorsOf(context)
-                                  .foreground100,
+                              color: ReownAppKitModalTheme.colorsOf(
+                                context,
+                              ).foreground100,
                               fontWeight: widget.linkMode
                                   ? FontWeight.bold
                                   : FontWeight.normal,
@@ -106,8 +106,9 @@ class SettingsPageState extends State<SettingsPage> {
                             'Socials Off',
                             textAlign: TextAlign.end,
                             style: TextStyle(
-                              color: ReownAppKitModalTheme.colorsOf(context)
-                                  .foreground100,
+                              color: ReownAppKitModalTheme.colorsOf(
+                                context,
+                              ).foreground100,
                               fontWeight: !widget.socials
                                   ? FontWeight.bold
                                   : FontWeight.normal,
@@ -134,8 +135,9 @@ class SettingsPageState extends State<SettingsPage> {
                           child: Text(
                             'Socials On',
                             style: TextStyle(
-                              color: ReownAppKitModalTheme.colorsOf(context)
-                                  .foreground100,
+                              color: ReownAppKitModalTheme.colorsOf(
+                                context,
+                              ).foreground100,
                               fontWeight: widget.socials
                                   ? FontWeight.bold
                                   : FontWeight.normal,
@@ -152,8 +154,9 @@ class SettingsPageState extends State<SettingsPage> {
                             'Analytics Off',
                             textAlign: TextAlign.end,
                             style: TextStyle(
-                              color: ReownAppKitModalTheme.colorsOf(context)
-                                  .foreground100,
+                              color: ReownAppKitModalTheme.colorsOf(
+                                context,
+                              ).foreground100,
                               fontWeight: !widget.analytics
                                   ? FontWeight.bold
                                   : FontWeight.normal,
@@ -180,8 +183,9 @@ class SettingsPageState extends State<SettingsPage> {
                           child: Text(
                             'Analytics On',
                             style: TextStyle(
-                              color: ReownAppKitModalTheme.colorsOf(context)
-                                  .foreground100,
+                              color: ReownAppKitModalTheme.colorsOf(
+                                context,
+                              ).foreground100,
                               fontWeight: widget.analytics
                                   ? FontWeight.bold
                                   : FontWeight.normal,
@@ -210,12 +214,21 @@ class SettingsPageState extends State<SettingsPage> {
                         title: 'Clear storage',
                         size: BaseButtonSize.small,
                         onTap: () async {
+                          final prefs = await SharedPreferences.getInstance();
+                          final keys = prefs.getKeys();
+                          for (var key in keys) {
+                            if (key.startsWith('appkit_sample_')) {
+                              await prefs.remove(key);
+                            }
+                          }
                           await widget.appKitModal.appKit!.core.storage
                               .deleteAll();
-                          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                            content: Text('Storage cleared'),
-                            duration: Duration(seconds: 1),
-                          ));
+                          ScaffoldMessenger.of(context).showSnackBar(
+                            SnackBar(
+                              content: Text('Storage cleared'),
+                              duration: Duration(seconds: 1),
+                            ),
+                          );
                         },
                       ),
                     ),
@@ -286,10 +299,7 @@ class SettingsPageState extends State<SettingsPage> {
                   children: [
                     Text('Bundle ID: ', style: textStyle),
                     Expanded(
-                      child: Text(
-                        snapshot.data ?? '',
-                        style: textStyleBold,
-                      ),
+                      child: Text(snapshot.data ?? '', style: textStyleBold),
                     ),
                   ],
                 );
@@ -303,10 +313,7 @@ class SettingsPageState extends State<SettingsPage> {
                   children: [
                     Text('Client ID: ', style: textStyle),
                     Expanded(
-                      child: Text(
-                        snapshot.data ?? '',
-                        style: textStyleBold,
-                      ),
+                      child: Text(snapshot.data ?? '', style: textStyleBold),
                     ),
                   ],
                 );